Avoid problems

+++ examples/plugin/source/main.cpp
@@ -284,10 +284,10 @@ struct MyPlugin : public yup::AudioProcessor
                         continue;
 
                     float volume = yup::jlimit (0.0f, 1.0f, getParameter (P_VOLUME).getValue() + 0.0f); // parameterOffsets[P_VOLUME]);
-                    sum += std::sinf (voice->phase * 2.0f * 3.14159f) * 0.2f * volume;
+                    sum += std::sin (voice->phase * 2.0f * 3.14159f) * 0.2f * volume;
 
-                    voice->phase += 440.0f * std::exp2f ((voice->key - 57.0f) / 12.0f) / sampleRate;
-                    voice->phase -= std::floorf (voice->phase);
+                    voice->phase += 440.0f * std::exp2 ((voice->key - 57.0f) / 12.0f) / sampleRate;
+                    voice->phase -= std::floor (voice->phase);
                 }
 
                 *outputL++ = sum;
